Add code for parallel execition of executors

+++ executor/executor.py
@@ -1,62 +1,61 @@
-# Позже здесь будет функциональность для запуска тасок на удаленных машинах
-import abc
+import time
 import subprocess
 from typing import Tuple, Optional
 import threading
 
 from models.status import Status
+from concurrent.futures import Future, CancelledError
 
 from paramiko.client import SSHClient
 
 
-class Executor(abc.ABC):
-    @staticmethod
-    @abc.abstractmethod
-    def run_stage(stage_command: str) -> Tuple[str, str]:
-        """
-        Execute a command
-        return stdout, stderr
-        """
-        pass
-
-
-class LocalExecutor(threading.Thread, Executor):
-    def __init__(self, *args, **kwargs):
+class LocalExecutor(threading.Thread):
+    def __init__(self, command: str, *args, **kwargs):
         super().__init__(*args, **kwargs)
         self._process: Optional[subprocess.Popen] = None
+        self._command: str = command
+        self._future = Future()
         self._status: Status = Status.PENDING
-        self._current_command = None
-        self._stdout = None
-        self._stderr = None
-
-    def run(self) -> None:
-        process = subprocess.Popen(self._current_command, stdout=subprocess.PIPE, stderr=subprocess.PIPE, shell=True, text=True)
-        self._status = Status.RUNNING
-        self._process = process
-        stdout, stderr = process.communicate()
-        if stderr:
-            self._status = Status.FAILED
-        else:
-            self._status = Status.SUCCESS
-        self._stderr = stderr
-        self._stdout = stdout
-
-    def set_command(self, command: str):
-        self._current_command = command
-
-    def get_result(self) -> Tuple[str, str]:
-        return self._stderr, self._stdout
 
     @property
-    def status(self) -> Status:
+    def status(self):
         return self._status
 
-    def stop_stage(self):
-        self._process.kill()
-        self._status = Status.FAILED
+    def run(self) -> None:
+        self._status = Status.RUNNING
+        try:
+            process = subprocess.Popen(self._command, stdout=subprocess.PIPE, stderr=subprocess.PIPE, shell=True,
+                                       text=True, encoding="cp866")
+            self._process = process
+            stdout, stderr = process.communicate()
+            if self._future.cancelled():
+                return
+
+            self._future.set_result((stdout, stderr))
+            if not stderr:
+                self._status = Status.SUCCESS
+            else:
+                self._status = Status.FAILED
+        except Exception as err:
+            self._future.set_exception(err)
+            self._status = Status.FAILED
+
+    def get_result(self, timeout: Optional[int] = None) -> Tuple[str, str]:
+        try:
+            return self._future.result(timeout)
+        except CancelledError:
+            return "", ""
+
+    def cancel(self):
+        if self._process is not None:
+            self._process.kill()
+            self._future.cancel()
+            self._status = Status.ABORTED
+        else:
+            raise ChildProcessError("Process is not created yet.")
 
 
-class SshExecutor(Executor):
+class SshExecutor(threading.Thread):
     @staticmethod
     def run_stage(stage_command: str) -> Tuple[str, str]:
         with SSHClient() as client:
@@ -66,3 +65,13 @@ class SshExecutor(Executor):
             _, stdout, stderr = client.exec_command(stage_command)
             return stdout.read().decode("utf-8"), stderr.read().decode("utf-8")
 
+
+if __name__ == '__main__':
+    local_executor = LocalExecutor("ping localhost", threading.Semaphore(value=1))
+    local_executor.start()
+    # print(local_executor.get_result())
+    print(local_executor.status)
+    time.sleep(0.01)
+    # local_executor.cancel()
+    print(local_executor.get_result())
+    # print(local_executor.status)
